+++ src/core/geometry/vertex.h
@@ -1,13 +1,13 @@
 #pragma once
 
-#include <cstdlib>
+#include <cstdint>
 
 namespace butter {
 
 struct Vertex {
 
-    size_t refPoint = 0; // Reference point index
-    size_t primIndex = 0; // Which primitive this vertex belongs to
+    uint32_t refPoint = 0; // Reference point index
+    uint32_t primIndex = 0; // Which primitive this vertex belongs to
 
 };
